MINIFI-171  Dynamic Properties support for processors
Added support in Processor.cpp to both set and get dynamic properties.
The logic will first attempt to set the parsed Property from the .yml
file as a supported property. IF that parsed Property is not in the
whitelisted supported list AND if the “nifi.accept.dynamic.properties”
configuration is set to true then the Property will be added to a
Map<String, Property> of dynamic properties that can be accessed in a
manner similar to the supported properties from any processor logic
that extends Processor.cpp