MINIFI-193: Ensure safe UTF encoding
Since the C++ library is agnostic of UTF-8 we can safely write the bytes.
Since we won't be interpreting the UTF-8 code in the core library
we do not need any additional dependencies. Nor do we need to worry about
encoding beyond proper serialization and deserialization of the byte array.

This closes #68.

Signed-off-by: Aldrin Piri <aldrin@apache.org>