MINIFI-275 Bugfix for YAML Configs without Component IDs
This commit fixes a bug in which loading config YAML that did not
specify IDs (UUIDs) for components caused an exception. The logic
previously treated component IDs as required fields by loading them
without checking for existence. This commit updates the logic to
generate IDs when they are not specified in the YAML config. As part
of this fix, the logic for loading connections from YAML config was
modified to search by name or remote port id in the absence of source
id or remote id in the YAML config.

For fields that are required, useful error messages were added when
those fields are missing to assist users in self-diagnosing issues
related to invalid config files.

Some minor tweaks to the top level .gitignore are included with this
commit.

This closes #85.

Signed-off-by: Aldrin Piri <aldrin@apache.org>