MINIFI-294 Required vs. optional fields in config YAML
Long description of the changes in this commit can be found in the
JIRA for MINFI-294. Basically, better checks are now in place and
in the case of missing required fields, better error messaging.
This builds upon improvements made in MINIFI-275, and also
updates the config.yml exampe in the README.md file to match the
code.