MiNiFCPP-DOC: Updated Outdated Bootstrapping Info
The paragraph in the Getting Started: Bootstrapping section originally told the user to enter N to continue the build process, but with the updated menu guided bootstrap process displayed in the console, P is what one must enter to continue the build process. I updated that info. 

The previous menu guided bootstrap process was outdated and did not show the new features Bustache Support to SQL Support and the Build Options. So, I updated it to show that information.