Update H2oPspScoreBatches processor
This update includes passing "index=False" to "batch_scores_df.to_string(index=False)". By updating this line of code, we tell pandas to not include the DataFrame index when converting the DataFrame to a string. The reason for this update is that we don't want this extra column pandas tries to add to our predictions frame, we only want the predictions. Thus, later when the predictions get saved to a csv file, it will only include the predictions.