MINIFICPP-1177 Fix log rotation handling in TailFile
Instead of the dodgy logic of doing one file at a time in onTrigger
and trying to keep track using the persisted state, we follow the same
logic as NiFi now: only look at rolled files if the input file got
truncated, only look at new rotated files matching the pattern, and
stream all new content in one onTrigger call.

Also fixed some bugs in Multiple file mode and Yield detection logic,
as well as corrected some unit tests (when we checked the log output,
we previously included output from earlier stages of the test).

Support some previously unsupported NiFi properties:
- Recursive lookup
- Lookup frequency
- Rolling Filename Pattern

Change the default delimiter to \n (the previous default was to always
read to the current end of the file, even if it is in the middle of a
line).  Also include the delimiter in the flow file.  (As NiFi does.)