MINIFICPP-1202 - Fix unstable failing test for MinifiConcurrentQueue
It is a test ensuring that the consumers that put back failing elements to their queue can eventually get to read new elements (so the queue is FIFO). The test was interrupting the consumer by stopping the queue. However there was no guarantee that the
consumer already finished reading the data out by this time.