MINIFICPP-1262 Move independent tests into separate test cases
MINIFICPP-1262 Wait for a condition with timeout

In one place, we waited for a fixed 10 ms (which may be too little), in two other
places we waited for a condition indefinitely in a loop (which may take forever).
Now we always wait for a condition, but time out after 1 second.

If the condition is still not true after the timeout, the test will abort (since
the consumer thread will be destructed without join), but this is handled well
by the test runner.

MINIFICPP-1262 Make the consumers more robust

- Increase the retry count of timed waiting consumers from 6 times 3ms to 1000 times 1ms.
- Introduce limits to the number of iterations in all consumers which didn't have one except for
  * those using dequeueWait or consumeWait (as these are blocking calls), and
  * getSpinningReaddingDequeueConsumerThread, as that is supposed to spin until it finds all inputs.
- Make the retry code in consumers not using dequeueWait or consumeWait more uniform.
- Remove the internal result counting from getConsumeWaitConsumerThread as this is now done externally.
- Remove some unused variables.

Signed-off-by: Arpad Boda <aboda@apache.org>

This closes #817