MINIFICPP-1262 Wait for a condition with timeout
In one place, we waited for a fixed 10 ms (which may be too little), in two other
places we waited for a condition indefinitely in a loop (which may take forever).
Now we always wait for a condition, but time out after 1 second.

If the condition is still not true after the timeout, the test will abort (since
the consumer thread will be destructed without join), but this is handled well
by the test runner.