MINIFICPP-1104 Add SSL docker tests for PublishKafka and automate them in CI
MINIFICPP-1104 Fix exception handling of py::eval

If pybind11 exceptions are not handled in the scope of the GIL
it causes a segmentation fault. In out case it caused a segmentation
fault if the to be evaluated python processor could not be loaded (e.g
missing dependency in imports).

See https://github.com/pybind/pybind11/issues/1803

MINIFICPP-1104 Fix broken integration tests

- Remove Windows specific DNS name
- Change default scheduling strategy to TIMER_DRIVEN in tests to bypass
  incoming connection tests for EVENT_DRIVEN strategy
- Change message timeout to 12 seconds to fix configuration error of
  linger.ms being higher than message.timeout.ms

MINIFICPP-1104 Switch spotify/kafka image to newer kafka image

- Separate wurstmeister/zookeeper and wurstmeister/kafka images are used
  now instead of spotify/kafka image for SSL testing.

MINIFICPP-1104 - SSL docker tests for PublishKafka

MINIFICPP-1104 Add docker tests to github workflows

- Add docker tests to ci.yml
- Change permissions of test files as github actions can only run docker
  containers as root
- Change timeouts to adapt to github's environment

MINIFICPP-1104 Speed up kafka integration tests

In kafka tests the observer was not working correctly, ignoring the
given subdirectory so it did not validate the output correctly. The
correct validation was only done after the timeout was reached.
Additionally the observer was never restarted when used multiple times
and the observer thread failed on directory changes as well.

Signed-off-by: Arpad Boda <aboda@apache.org>

This closes #878