MINIFICPP-1104 Speed up kafka integration tests
In kafka tests the observer was not working correctly, ignoring the
given subdirectory so it did not validate the output correctly. The
correct validation was only done after the timeout was reached.
Additionally the observer was never restarted when used multiple times
and the observer thread failed on directory changes as well.