MINIFICPP-1360 Configurable scheduling strategy in integration tests
The scheduling strategy was hard coded to be TIMER_DRIVEN for all
processors, which is not desirable in all cases. The processors have been
investigated on a case by case basis and changed to the most appropriate
default value. The desired scheduling can now be passed as a parameter
for all processors.