MINIFICPP-1323 Give the build config to cpack
Multi configuration generators (visual studio) don't save the last build type in the cache,
so cpack apparently has no way of knowing that it has to get the binary from the Debug/RelWithDebInfo
directory and defaults to Release (the first one AFAIK). Not sure why the issue only happens with
install(TARGETS), but the fix is to append pass -C <configuration> to cpack. -- thanks to @szaszm