MINIFICPP-1391 Remove XCode 10.3 CI job, add 12.0
MINIFICPP-1391 Replace usages of set-env

set-env has been deprecated and triggers a warning; see
https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/

MINIFICPP-1391 Remove the CMAKE_BUILD_OPTIONS env setting

This environment setting doesn't do anything, and it gives the false
appearance as if the Pcap extension was enabled (it isn't).

As the Pcap extension is not in a good shape, I have left it disabled.

ENABLE_LUA_SCRIPTING seems to work (at least it compiles), so I enabled it
in the jobs where the intention was to enable it.

MINIFICPP-1391 Increase the time limit on the docker test

Signed-off-by: Arpad Boda <aboda@apache.org>

This closes #925