MINIFICPP-1391 Remove the CMAKE_BUILD_OPTIONS env setting
This environment setting doesn't do anything, and it gives the false
appearance as if the Pcap extension was enabled (it isn't).

As the Pcap extension is not in a good shape, I have left it disabled.

ENABLE_LUA_SCRIPTING seems to work (at least it compiles), so I enabled it
in the jobs where the intention was to enable it.