Stabilize last_write_time checks in FileUtilsTests
last_write_time using both boost and stat  could be off by a
millisecond which could cause the check to fail. Additional 10
milliseconds before writing the file makes sure the saved time and the
file's timestamp are in order.

On Windows we can experience that the last_write_time of the same file
can be off by a milliseconds between 2 call. Checking this should be
disabled on Windows.