Fix flakiness and speed up TailFileTests
Sometimes it could occur that the generated long lines contained the
same character string that we added in the next line and expected not to
be in the final payload because of the maximum payload limit. By
distinguishing the character set from which the generated lines and the
added lines are consisted of this can be eliminated.

Timeout from checks where we do not expect a log to be present was also
removed to speed up the tests.

One sleep was also missing between one instance of log controller reset
and temp file appending.