MINIFICPP-1585 Rename enum values in ConsumeKafkaTests::PublishEvent
TRANSACTION_COMMIT is defined in winnt.h as a macro, so the build fails
on Windows.  I have renamed it, and a few other enum values for
consistency.

Also added /K to the VS2019 Windows CI job, to make sure bugs like this
get caught in the CI stage in the future.