MINIFICPP-1619 Make sure serialized records expire in DBProvenanceRepositoryTests
DBProvenanceRepositoryTests transiently failed due to not waiting enough time to have the serialized records to expire in the test. The failure does not seem to be connected to a specific change as it seems to transiently fail with commits from a few months ago, but the failures seem to increase in the past week. Increasing the wait time from 1,5 to 2 seconds solves the issue.

https://issues.apache.org/jira/browse/MINIFICPP-1619

Closes #1155

Signed-off-by: Marton Szasz <szaszm@apache.org>