MINIFICPP-1134 disable scripting extension
This allows us to avoid linking to libpython by default, leaving only
system dependnecies like libc, c++ standard lib, pthread, libgcc, etc.
This in turn allows us to distribute only one convenience binary for all
targeted linux systems if we ensure that we're building on one with the
oldest glibc. That would be CentOS 7 with devtoolset-10.

Signed-off-by: Ferenc Gerlits <fgerlits@gmail.com>

This closes #1174