MINIFICPP-1658 Fix handling OS signals when running in container
If Minifi runs in a docker container signals are not handled properly because the starter shell script does not forward the signals to the spawned process. This could cause problems like being unable to stop a running container with SIGTERM signal, for example when shutting down a pod in Kubernetes environment, resulting in always waiting for the grace period to end, or locally with CTRL+C combination's SIGINT signal.
Instead of spawning a new process replacing the shell script with the Minifi process when starting it solves the issue.

Closes #1190
Signed-off-by: Marton Szasz <szaszm@apache.org>