MINIFICPP-1222 Remove engine queue from ExecutePythonProcessor
Originally the Jira ticket was about reowrking the engine queue to only allow the maximum concurrent tasks number of script engines to run at the same time. The problem with that was that even if the script engines were executed concurrently they use the same python interpreter with the global interpreter lock in place so no concurrency can take place.

The only possibility for concurrency to use sub-interpreters which would be also problematic. First of all pybind11 does not support an interface for sub-interpreters, it is only available through the CPython API, but the documentation also warns about the caveats of the sub-interpreters here: https://pybind11.readthedocs.io/en/stable/advanced/embedding.html
Secondly using the CPython API for sub-interpreters would require us to always know which thread we are running on and create separate sub-interpreters and thread states for each thread (in this case it would possibly not even allow us to run the python scripts from the main thread anymore). Managing the thread states and the sub-interpreters would be cumbersome in our architecture.

Because all of the above problems I removed the python script queue from the ExecutePythonProcessor which will run the python scripts on a single thread.

Closes #1227
Signed-off-by: Marton Szasz <szaszm@apache.org>