MINIFICPP-1710 MINIFICPP-1714 Stabilize docker tests
Recently some docker tests were failing randomly in the CI in the OPC and Kafka test suites. The root cause for both suite failures was the same that NiFi's and MiNiFi's PutFile processor creates a hidden temporary file before renaming it to the final version when all content has been written. The scheduling of the threads can cause that the temp file can be non-existent anymore when it is read or causes a premature notification of the file output creation. These files should be ignored in the notification and the output verification process.

This fix also revealed a flakiness of the ListS3 test due to a timing issue of the creation and the listing of S3 objects. The test was changed to be more stable.

https://issues.apache.org/jira/browse/MINIFICPP-1710
https://issues.apache.org/jira/browse/MINIFICPP-1714

Closes #1235
Signed-off-by: Marton Szasz <szaszm@apache.org>