MINIFICPP-1740 Add FetchFile processor
FetchFile added mostly according to NiFi's implementation. A difference
between this and NiFi's implementation is that the flow is only
transferred to failure due to completion strategy failure if it is
explicitly set to fail (in case of move strategy conflict). In NiFi's
case it's a bit inconsistent as it could also fail if the move target
directory could not be created or it has permission problems, but the
flow is transferred to success in any other completion strategy failure.

Closes #1262
Signed-off-by: Marton Szasz <szaszm@apache.org>