MINIFICPP-1788 flush old bins on expiration
Before this change, if there were old flow files in
MergeContent/BinFiles, they were stuck until there was another incoming
flow file to the processor, because the processor was not triggered,
regardless of scheduling strategy.

Now the processor is always triggered, regardless of incoming flow
files, so it can check and flush expired bins. This didn't result in a
noticably increased CPU usage even with EVENT_TRIGGERED scheduling
strategy on my system.