increase C2PropertiesUpdateTests heartbeat period
So that the consumer thread can eventually process all of the operations
and doesn't get overwhelmed. My suspicion is that the test failure on
windows occurred because there was always at least one request in the
C2Agent queue, therefore it never reached the point of being able to
restart without losing data.