MINIFICPP-1675 Use regex.h instead of std::regex when using libstdc++
Due to a bug in the libstdc++ implementation of std::regex it is unsafe
to use std::regex_match and std::regex_search on large texts as the
process can crash due to a stack overflow. In this commit std::regex
usages are replaced with minifi::utils::Regex which uses regex.h if
libstdc++ is used otherwise it uses std::regex.

There are a few exceptions:
- std::regex is still used for regex_replace calls
- AppendHostInfo still uses std::regex as POSIX regex grammar is not
  sufficient for easily filtering interface names, and there is no risk
  of running into the bug as the interface names are small
- In tests and mocks we still use std::regex as they are only run in our
  test environments with controlled inputs and sometimes use regex
  patterns not available in POSIX grammar

Closes #1300
Signed-off-by: Marton Szasz <szaszm@apache.org>