MINIFICPP-1828 fix InvokeHTTP Attributes to Send
It was using subsequence matching (search) instead of full matching
(match).

Additionally, renamed "pattern" to "string" in RegexUtils, because it
was used to refer to the string in which we looked for the pattern
(regex). It was confusing.

unrelated change: added processor name to backpressure logging to help
debugging

add utils::filter to filter optional values