MINIFICPP-1861 Fix cmake command generation in bootstrap.sh
The bootstrap script did not account for the enabled by default cmake
options. Instead, it assumed that the ENABLE_* options are all OFF by
default and only generated the cmake flags for those options if they
were turned on explicitly. Because of this even when an option was set
to disabled in the bootstrap script, it was built with the generated
cmake command if that feature was enabled by default in cmake.

Closes #1351
Signed-off-by: Marton Szasz <szaszm@apache.org>