MINIFICPP-1888 Move all extension builds to CentOS job
With more extensions added to be turned on by default, the centos build
targeting the default extensions takes longer to finish. To make sure we
can build on both ubuntu/gcc and CentOS and to minimize the CI resources
used, all extension builds that are available in the CentOS docker build
are moved to the CentOS job and tests are run on CentOS and in the clang
Ubuntu build.

Closes #1374
Signed-off-by: Marton Szasz <szaszm@apache.org>