MINIFICPP-1906 Restore original timeout for CivetServer
There was a change in CivetWeb (commit hash: a7361235) where a
previously used poll function with constant timeout of 200 ms was
replaced with the mg_poll function. This function uses the
SOCKET_TIMEOUT_QUANTUM constant for the timeout value of a single poll
call. The default value of this constant is 2000 ms and due to this the
shutdown of the CivetServer had to wait 10 times more for the ongoing
poll to finish in every single test case, which increased the runtime of
the test cases. The SOCKET_TIMEOUT_QUANTUM is changed back to 200 ms in
this PR.

Closes #1403
Signed-off-by: Marton Szasz <szaszm@apache.org>