MINIFICPP-1921 Change test server ports residing in the default range
Two of the socket tests randomly fails sometimes with the error message
`bind: Address already in use`: the SecureSocketGetTCPTestEmptyPass and
the TLSServerSocketSupportedProtocolsTest tests. One of these tests uses
the port 38776 and the other uses port 38778. The common issue with
these two ports is that both are in the default linux port range which
is between 32768 and 60999. If other tests are run beside these two
tests it is possible that the system assigns these two ports to the
clients in other tests, thus when we try to initialize the servers, the
port binding fails. We should change these two ports to be outside the
default system port range to avoid this issue.

Closes #1406
Signed-off-by: Marton Szasz <szaszm@apache.org>