MINIFICPP-1947 add EXCLUDE_FROM_DEFAULT_BUILD to linter
Currently we build the whole solution on windows which linters are part
of. MINIFICPP-1575 intruduced an ENABLE_LINTER cmake option to disable
linter, so it can be built on windows without python. But on the CI we
currently run the linters twice (once implicitly in the build step, and
once again explicitly by the linter step, this also means a linter
failure shows up as a build failure in the CI which can be confusing at
a glance)

If we set the EXCLUDE_FROM_DEFAULT_BUILD target option on all linter
targets, the linters wont be built implicitly by the build step, this
also means we can rid of the ENABLE_LINTER cmake option.

Closes #1427
Signed-off-by: Marton Szasz <szaszm@apache.org>