MINIFICPP-2024 Refactor repository hierarchy
We would like to handle all repositories the same way for reporting
metrics in MINIFICPP-2022 but the current hierarchy structure does not
allow that to be done easily.

The inheritance hierarchy of repositories is unnecessarily too complex.

While the base of the flow repositories is core::Repository and the base
of content repositories is core::ContentRepository, the
VolatileContentRepository inherits from both that does not seem logical.

The VolatileRepository is a template class that only works with two
specific set of template argument and has unnecessary implementations of
base classes that is not valid for all child volatile repositories.

There is also a virtual inheritance from core::SerializableComponent for
flow repositories that should not be needed to be virtual.

This hierarchy should be cleaned up to be able to create a single
interface for all repositories for metric reporting.

Closes #1485
Signed-off-by: Marton Szasz <szaszm@apache.org>