MINIFICPP-2034 Cache SID lookups in CWEL
The SID -> Username lookup is a system call, which can be quite slow. As
this mapping is unlikely to change without a Windows (and so MiNiFi)
restart, we should cache it.

Caching is quite simple: there is a hard-coded 24 hour cache expiry, the
cache is in memory only, and the only way to clear it is to restart
MiNiFi. We can improve this later if a user asks for it.

Closes #1502
Signed-off-by: Marton Szasz <szaszm@apache.org>