change how the Path Filter property works to match NiFi
Before this change, files in the input root directory were always listed.
After the change, the files in the input root directory are only listed if
the value of the Path Filter property matches ".".

Also to match NiFi's behavior, we now ignore the Path Filter property if
Recurse Subdirectories is false.