MINIFICPP-2053 Reduce CentOS image size in CI build
It started to occur that the CentOS build exceeded the 14GB disk space
Github Actions VMs provide (for example
https://github.com/apache/nifi-minifi-cpp/actions/runs/4263514879/jobs/7420351598).
JDK and the packaged tar.gz are removed from the CI build and JNI and
SFTP builds are moved to the ubuntu build. This reduces the image size
by about 2GBs.

Closes #1513
Signed-off-by: Marton Szasz <szaszm@apache.org>