MINIFICPP-2170 Fix system certificate store usage in SSLContextService on Linux
Use utils::getDefaultCAFile() in the SSLContextService to find the system CA
certificate store file, as the compiled-in default value is usually wrong.
If the system CA certificate store file is still not found, you can set the
SSL_CERT_DIR or SSL_CERT_FILE environment variables to tell OpenSSL where
the certificates are.

Also: fix the incorrect handling of InvokeHTTP::DisablePeerVerification.
Signed-off-by: Gabor Gyimesi <gamezbird@gmail.com>

This closes #1620