MINIFICPP-2134 Do not swallow parsing exceptions
If the value of a typed property (Time Period, Data Size, Port etc) is set,
but to an invalid value, then we should throw in getProperty<T>().

In some cases, we were using a typed getProperty<T>(), but the corresponding
PropertyDefinition did not have a PropertyType, so the C2 server had no way
of validating the property.  In these cases, I have kept the current behavior,
which is to treat the invalid value as if no value was set -- but I have also
added the PropertyType to the PropertyDefinition so we can add validation in
a later release.

Additionally, the Max Poll Records property of ConsumeKafka used to internally
treat the value as uint64, but in the manifest we sent a uint32 validator;
I have changed the validator to be uint64 ("unsigned long"), as well.

Closes #1610

Signed-off-by: Marton Szasz <szaszm@apache.org>