MINIFICPP-2181 Use systemd service management on Linux
I generated small parts of this PR with GitHub Copilot. I believe all of the
generated snippets are trivial enough to not be copyrightable in their own
right, and I would have written the same myself. For this reason, I believe
that I own the copyright for the whole code, and it can be licensed under the
Apache License 2.0.

Signed-off-by: Ferenc Gerlits <fgerlits@gmail.com>
This closes #1627