MINIFICPP-2290 Support the default schema when serializing a JSON config
The JSON flow configuration comes in two flavors: the default schema and
the NiFi schema. In the default schema, the root node location is ".",
which is a notation borrowed from yaml-cpp to refer to the current node
(so node["."] == node).

This notation is not supported directly by rapidjson; it works when
parsing the JSON input because it is implemented in
flow::Node::operator[].

We need to handle node["."] when writing the JSON flow configuration,
too.

Closes #1719

Signed-off-by: Marton Szasz <szaszm@apache.org>