MINIFICPP-2325 Add missing compiler flags
These flags are on by default when we use msbuild, but they need to be
added manually when we build using ninja. Without these flags, the
resulting dlls were roughly twice as large when using ninja compared to
using msbuild. After adding the flags, the sizes are back to normal.