MINIFICPP-2325 Add missing compiler flags
These flags are on by default when we use msbuild, but they need to be
added manually when we build using ninja. Without these flags, the
resulting dlls were roughly twice as large when using ninja compared to
using msbuild. After adding the flags, the sizes are back to normal.

Additionally, on Windows, always copy minifi.exe to the minifi_main directory

This was previously done in prep_for_win_package, but the location of
minifi.exe depends on the generator (Ninja or Msbuild) and the build
type, and this info is not easily available at packaging time.

Closes #1755

Signed-off-by: Marton Szasz <szaszm@apache.org>