Verified Can Create MiNiFi Conan Package after removing MINIFI_BUILD_CONAN_PACKAGE
One of the concerns was that we wanted to make sure that README.md, LICENSE
, NOTICE and even the binary ones were included in the MiNiFi C++ conan package.

I removed the MINIFI_BUILD_CONAN_PACKAGE condition, so CPACK_PACKAGE_DESCRIPTION_FILE
and CPACK_RESOURCE_FILE_LICENSE are not ignored when creating the MiNiFi conan package.
This is for README.md and LICENSE.txt and NOTICE files including the binary ones.

As an alternative to using CPACK with CMake to make sure these files are included
in the MiNiFi conan package, through conanfile.py we assign our shared_sources variable
to export_sources and in our py file's package() method, we run cmake.install() to
ensure that the source files and binary files assets are copied into the MiNiFi conan package.

Also for the MiNiFi conan package, we copy over the .h, .i, .a and .so files into their
appropriate include/ folder and lib/ folder.