Refactored MiNiFi Creating Properties via to_array to Dynamically Dtrm Num Properties
Across Processors and ControllerServices, we previously determined the number of properties by manually
specifying it to create the static properties using std::array. With this update, for properties that consist
of more than 0, we switch to using C++20's std::to_array<>({prop1, prop2, etc}) to automatically determine
the number of properties using list comprehension. Thus, for MiNiFi Processors and ControllerServices that
had at least 1 property or more, we used this std::to_array approach to do the conversion of PropertyDefinition
objects to PropertyReference objects. In the case that we had Processors or ControllerServices that had 0
properties, we still needed to use the previous approach of manually specifying that there were 0 properties.