Created RocksDB Conan Recipe to take patches like BundledRocksDB
Also when we create the RocksDB conan package, it not only builds with the patches like BundledRocksDB, it also
incorporates ZLib and ZStd conan packages too.

Therefore, now that we use a conan RocksDB for building MiNiFi with conan, when we build MiNiFi with standalone
CMake, we use the original version of BundledRocksDB that was in 'main' branch before I modified it.

I verified that when I build MiNiFi with conan, 218 CTESTs pass out of 219.
Similarly, when I build MiNiFi with standalone, then I run CTESTs, 218 out of 219 pass too.