MINIFICPP-2487 - Destroy PyException while holding GIL
PyException holds a reference to the python exception object. When it escapes
the scope of the GIL (into an outer catch block), its destruction crashes the
application, and the python library logs the reason as not holding the lock.
This can occur when e.g. the describe function is malformed (declared with the
incorrect number of positional arguments).

Signed-off-by: Ferenc Gerlits <fgerlits@gmail.com>
This closes #1891