MINFICPP-2243 ListenHTTP: process requests in onTrigger
Network calls are now only made on within onTrigger, instead of reading
data on a background thread and storing it in memory. The incoming data
is stored in the kernel buffers before we read it out.

With these changes, we avoid keeping incoming data in memory, and we can
avoid creating flow files without a normal onTrigger session.

Closes #1826 "In ListenHTTP process incoming request only in onTrigger"

Signed-off-by: Marton Szasz <szaszm@apache.org>