MINIFICPP-2461 Fix linker error on MSBuild
On Windows, if we use MSBuild instead of Ninja, binary artifacts are put
in a subdirectory according to the build type, eg. bin/Debug or bin/Release.
Since our LibArchive build was not aware of this, it could not find the
liblzma.lib file. This change is a workaround for this problem.