MINIFICPP-2544 extend and fix FIPS docs
Added a local README in the subdirectory, to prevent user confusion, especially with the binary package: If a user doesn't want FIPS modules, they might be surprised to see a `fips` subdirectory in their installation and go investigate. Adding a message and reassuring them that it's unused until they explicitly enable it could prevent user confusion and build goodwill.

Closes #1948

Signed-off-by: Martin Zink <martinzink@apache.org>