MINIFICPP-2636 update ExecutePythonProcessor description
- onTrigger accepts session, not property
- justifying basic behavior with efficiency seems weird, removed
- remove the sentence about concurrency, since the processor is single threaded

On a somewhat unrelated note, should we drop this API in favor of the newer NiFi-compatible python scripting API, or is it still useful to maintain? I guess it's easier to write processors for the legacy API, but the new API seems better designed, plus it handles dependencies. If we drop it, it should be done before 1.0.

Closes #2035

Signed-off-by: Martin Zink <martinzink@apache.org>