MINIFICPP-2642 std::invocable should be used with std::invoke
Replacing calls with invoke where invocable is used.

https://en.cppreference.com/w/cpp/concepts/invocable
The invocable concept checks that type F can be called with arguments
Args... using std::invoke. Std::invoke is more general than a direct
function call: it supports member function and data member pointers too,
where the first "argument" is the object.

In theory, if we checked for invocable but called directly, someone
could pass a member pointer as the callback, and it would pass the
concept check but fail to build.

Closes #2038

Signed-off-by: Marton Szasz <szaszm@apache.org>