MINIFICPP-2624 Modularized behave tests part 1
The docker tests are a great way to do integration testing but it became a bit
clunky over the years.  I've recreated the framework with simplicity in mind.
Its now be able to create a flow (without concrete processor classes) and do
simple checking with them. Any domain specific checking or containers should be
moved into the respective extension.

This will enable us to use the framework to test separetly built extensions in
the future (upcoming stable c api).  And hopefully this will make writing of
these tests to be a bit more straight forward.

Signed-off-by: Ferenc Gerlits <fgerlits@gmail.com>

Closes #2044